requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
idna==3.6
livekit==0.9.1
livekit-api==0.4.2
livekit-protocol==0.3.1
multidict==6.0.5
numpy==1.26.4
opencv-python==4.9.0.80
protobuf==4.25.3
PyJWT==2.8.0
python-dotenv==1.0.1
types-protobuf==4.24.0.20240302
yarl==1.9.4
